Another main, application and listener implementation
This implementation doesn't subclass StreamPlayer, and it doesn't implement the StreamPlayerListener interface. Instead, the application has an instance of the regular StreamPlayer, and there is a separate implementation of the StreamPlayerListener interface.

One thing I dislike, is that the StreamPlayerListener depends on the StreamPlayer; it imports it, and has an instance of it, because that is how it gets the total number of bytes.

I have also added an alternative progress logging message.

I would have prefered to create the listener i main(), but it needs information that only the application has (the filename), so it has to be created in the application.

Besides that, it works as usual.