Some unit tests
In the TimeTool test, I think I've found a bug. Check lines 45 and 46.

The tests are written to document the current behavior, rather than to guard against bugs.

The time duration test don't test very much. It's because the method needs input data that I can't easily create in the test. With test-driven development, the method would have been written in a different way.