Constructors with parameters.
Two new constructors are provided. They allow the user to replace the default logger and executor services.

Constructor chaining is used. The parameterless constuctor now calls the logger constructor, which calls the construtor with executor arguments. In this way, by using and testing the parameterless constructor, we use and test all three constructors. (It's tested and it works).

A use case for the new constructors are:
Unit testing without having to see a lot of logging messages)
Unit testing where we spy on the executors, using a mocking framework, such as Mockito.

An advantage from an architecture point of view, is that depencencies become more visible. A future refactoring could be to remove the parameterless constructor. Then this class wouldn't be dependent on a logger implementation, but only on the Logger interface. Same with the executor services.