Mockito + demo
Mockito is a mocking library used to create faked instances of classes. The class doesn't need to exist yet; we can create a mock from an interface or a parent class.

Mockito can also create spies, which intercept existing methods and check if and how they are called. The demonstration unit test uses spies.

Mockito has JUnit as a dependency. To avaoid JUnit version number conflict, we back the Junit version from 5.5.1 to 5.1.1, which is what this verison of Mockito wants.