Test with reset() behaves oddly
reset() is a public method. If it's called when the player is playing, it seems to not return. I cannot really see what is executing when I run the test in the debugger.

The test print "play()", and goes on to play the song. It never prints "reset()"

Expected behavior: The player is stopped so quickly, that no music is heard. Maxium one frame of sound samples.
The word "reset() should be printed after "play()".

What can be learnt from this behavior?