Removed method StreamPlayer.getPositionByte()
The method was used in a unit test, as a means to
test the StreamPlayer.seekTo() method, but it never
worked. The StreamPlayer.getPositionByte() didn't seem
to work.

There are no other callers of getPositionByte() method in this codebase.
It is not used by XR3Player. It seems safe to remove it.

If the method really worked (which I doubt), then this is a
non-backwards-compatible change.

This change is a part of an effort to make all unit tests pass, by either updating the production code, the tests, or remove the failing tests.